replace band and tour with title and subtitle
